sandali_bot/requirements.txt
# Core bot framework
aiogram==3.4.1

# Scheduling
APScheduler==3.10.4

# Data and database
aiosqlite==0.19.0

# Visualization
matplotlib==3.8.4
numpy==1.26.4
seaborn==0.13.2

# Dotenv for config/secrets
python-dotenv==1.0.1

# HTTP requests for API calls (CoinGecko, Alpha Vantage, etc.)
requests==2.28.0
httpx==0.27.0

# Optional: LLM and NLP (DeepSeek etc., placeholder for now)
openai==1.30.1

# Other useful dev tools
loguru==0.7.2
